Unified setup.py files
